backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
pgvector==0.3.5
anthropic==0.39.0
google-cloud-vision==3.7.4
google-cloud-storage==2.18.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
slowapi==0.1.9
PyPDF2==3.0.1
tiktoken==0.8.0
psycopg2-binary==2.9.9